requirements.txt
# Core Teleg# Core Bot
Telethon==1.37.0

# AI Integration
anthropic==0.40.0

# HTTP & Async
aiohttp==3.11.10

# Hotel API
amadeus==8.1.0

# Environment & Utils
python-dotenv==1.0.1
Pillow==11.0.0
requests==2.32.3ing (optional, for future features)
Pillow==11.0.0

# Additional utilities
requests==2.32.3